Accept empty memos for sends; More tests
