+++ counterparty-core/counterpartycore/test/regtest/scenarios/scenario_19_mpma.py
@@ -1,0 +1,628 @@
+from binascii import hexlify
+
+
+def to_hex(x):
+    return hexlify(x.encode()).decode()
+
+
+SCENARIO = [
+    {
+        "title": "Create asset MPMASSET",
+        "transaction": "issuance",
+        "source": "$ADDRESS_1",
+        "params": {
+            "asset": "MPMASSET",
+            "quantity": 1000 * 10**8,
+            "divisible": True,
+            "description": "My super asset B",
+        },
+    },
+    {
+        "title": "Send XCP",
+        "transaction": "send",
+        "source": "$ADDRESS_1",
+        "params": {
+            "asset": "MPMASSET",
+            "quantity": 1000,
+            "destination": "$ADDRESS_2",
+        },
+    },
+    {
+        "title": "Send MPMA with memo",
+        "transaction": "mpma",
+        "source": "$ADDRESS_1",
+        "params": {
+            "assets": "XCP,MPMASSET,MPMASSET",
+            "quantities": "10,10,10",
+            "destinations": "$ADDRESS_2,$ADDRESS_3,$ADDRESS_4",
+            "memo": "the memo",
+        },
+        "controls": [
+            {
+                "url": "blocks/$BLOCK_INDEX/events?event_name=MPMA_SEND,CREDIT,DEBIT",
+                "result": [
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_10",
+                        "params": {
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_2",
+                            "memo": "the memo",
+                            "msg_index": 2,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_9",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_4",
+                            "memo": "the memo",
+                            "msg_index": 1,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_8",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_3",
+                            "memo": "the memo",
+                            "msg_index": 0,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_7",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_1",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_6",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_1",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 20,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_5",
+                        "params": {
+                            "address": "$ADDRESS_2",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_4",
+                        "params": {
+                            "address": "$ADDRESS_4",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_3",
+                        "params": {
+                            "address": "$ADDRESS_3",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                ],
+            },
+        ],
+    },
+    {
+        "title": "Send MPMA with memos for each send",
+        "transaction": "mpma",
+        "source": "$ADDRESS_1",
+        "params": {
+            "assets": "XCP,MPMASSET,MPMASSET",
+            "quantities": "10,10,10",
+            "destinations": "$ADDRESS_2,$ADDRESS_3,$ADDRESS_4",
+            "memos": ["memo1", "memo2", "memo3"],
+        },
+        "controls": [
+            {
+                "url": "blocks/$BLOCK_INDEX/events?event_name=MPMA_SEND,CREDIT,DEBIT",
+                "result": [
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_10",
+                        "params": {
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_2",
+                            "memo": "memo1",
+                            "msg_index": 2,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_9",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_4",
+                            "memo": "memo3",
+                            "msg_index": 1,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_8",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_3",
+                            "memo": "memo2",
+                            "msg_index": 0,
+                            "quantity": 10,
+                            "source": "$ADDRESS_1",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_7",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_1",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_6",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_1",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 20,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_5",
+                        "params": {
+                            "address": "$ADDRESS_2",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_4",
+                        "params": {
+                            "address": "$ADDRESS_4",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_3",
+                        "params": {
+                            "address": "$ADDRESS_3",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                ],
+            },
+        ],
+    },
+    {
+        "title": "Send MPMA with hex memo",
+        "transaction": "mpma",
+        "source": "$ADDRESS_2",
+        "params": {
+            "assets": "XCP,MPMASSET,MPMASSET",
+            "quantities": "10,10,10",
+            "destinations": "$ADDRESS_1,$ADDRESS_3,$ADDRESS_4",
+            "memo": to_hex("the memo"),
+            "memo_is_hex": True,
+        },
+        "controls": [
+            {
+                "url": "blocks/$BLOCK_INDEX/events?event_name=MPMA_SEND,CREDIT,DEBIT",
+                "result": [
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_10",
+                        "params": {
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_1",
+                            "memo": to_hex("the memo"),  # bytes hexified by the API
+                            "msg_index": 2,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_9",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_4",
+                            "memo": to_hex("the memo"),
+                            "msg_index": 1,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_8",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_3",
+                            "memo": to_hex("the memo"),
+                            "msg_index": 0,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_7",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_2",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_6",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_2",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 20,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_5",
+                        "params": {
+                            "address": "$ADDRESS_1",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_4",
+                        "params": {
+                            "address": "$ADDRESS_4",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_3",
+                        "params": {
+                            "address": "$ADDRESS_3",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                ],
+            },
+        ],
+    },
+    {
+        "title": "Send MPMA with memos and memo",
+        "transaction": "mpma",
+        "source": "$ADDRESS_2",
+        "params": {
+            "assets": "XCP,MPMASSET,MPMASSET",
+            "quantities": "10,10,10",
+            "destinations": "$ADDRESS_1,$ADDRESS_3,$ADDRESS_4",
+            "memos": ["memo1", "", "memo3"],
+            "memo": "the memo",
+        },
+        "controls": [
+            {
+                "url": "blocks/$BLOCK_INDEX/events?event_name=MPMA_SEND,CREDIT,DEBIT",
+                "result": [
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_10",
+                        "params": {
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_1",
+                            "memo": "memo1",
+                            "msg_index": 2,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_9",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_4",
+                            "memo": "memo3",
+                            "msg_index": 1,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "MPMA_SEND",
+                        "event_index": "$EVENT_INDEX_8",
+                        "params": {
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "destination": "$ADDRESS_3",
+                            "memo": "the memo",
+                            "msg_index": 0,
+                            "quantity": 10,
+                            "source": "$ADDRESS_2",
+                            "status": "valid",
+                            "tx_hash": "$TX_HASH",
+                            "tx_index": "$TX_INDEX",
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_7",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_2",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "DEBIT",
+                        "event_index": "$EVENT_INDEX_6",
+                        "params": {
+                            "action": "mpma send",
+                            "address": "$ADDRESS_2",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "event": "$TX_HASH",
+                            "quantity": 20,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_5",
+                        "params": {
+                            "address": "$ADDRESS_1",
+                            "asset": "XCP",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_4",
+                        "params": {
+                            "address": "$ADDRESS_4",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                    {
+                        "event": "CREDIT",
+                        "event_index": "$EVENT_INDEX_3",
+                        "params": {
+                            "address": "$ADDRESS_3",
+                            "asset": "MPMASSET",
+                            "block_index": "$BLOCK_INDEX",
+                            "calling_function": "mpma send",
+                            "event": "$TX_HASH",
+                            "quantity": 10,
+                            "tx_index": "$TX_INDEX",
+                            "utxo": None,
+                            "utxo_address": None,
+                        },
+                        "tx_hash": "$TX_HASH",
+                    },
+                ],
+            },
+        ],
+    },
+]
